MB-35 savings

+++ app/src/Service/BudgetService.php
@@ -6,15 +6,23 @@ namespace App\Service;
 
 use App\Dto\Budget\Http\BudgetFilterQuery;
 use App\Dto\Budget\Payload\BudgetPayload;
+use App\Dto\Budget\Response\BudgetResponse;
+use App\Dto\Budget\Response\ExpenseResponse;
+use App\Dto\Budget\Response\IncomeResponse;
 use App\Entity\Budget;
 use App\Entity\User;
+use App\Enum\ErrorMessagesEnum;
 use App\Repository\BudgetRepository;
+use App\Security\Voter\BudgetVoter;
+use Carbon\Carbon;
 use Doctrine\Common\Collections\Criteria;
-use Knp\Bundle\PaginatorBundle\Pagination\SlidingPagination;
+use My\RestBundle\Dto\PaginatedResponseDto;
+use My\RestBundle\Dto\PaginationMetaDto;
 use My\RestBundle\Dto\PaginationQueryParams;
 use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 class BudgetService
 {
@@ -23,32 +31,37 @@ class BudgetService
         private readonly IncomeService $incomeService,
         private readonly ExpenseService $expenseService,
         private readonly Security $security,
+        private readonly AuthorizationCheckerInterface $authorizationChecker,
     ) {
     }
 
-    public function get(int $id): Budget
+    public function get(int $id): BudgetResponse
     {
         $budget = $this->budgetRepository->find($id);
 
         if ($budget === null) {
-            throw new NotFoundHttpException('Budget not found');
+            throw new NotFoundHttpException(ErrorMessagesEnum::BUDGET_NOT_FOUND->value);
         }
 
-        $this->checkAccess($budget);
+        if (! $this->authorizationChecker->isGranted(BudgetVoter::VIEW, $budget)) {
+            throw new AccessDeniedHttpException(ErrorMessagesEnum::ACCESS_DENIED->value);
+        }
 
-        return $budget;
+        return $this->createBudgetResponse($budget);
     }
 
-    public function create(BudgetPayload $budgetPayload): Budget
+    public function create(BudgetPayload $budgetPayload): BudgetResponse
     {
         $budget = new Budget();
 
         return $this->createOrUpdateBudget($budgetPayload, $budget);
     }
 
-    public function update(BudgetPayload $budgetPayload, Budget $budget): Budget
+    public function update(BudgetPayload $budgetPayload, Budget $budget): BudgetResponse
     {
-        $this->checkAccess($budget);
+        if (! $this->authorizationChecker->isGranted(BudgetVoter::EDIT, $budget)) {
+            throw new AccessDeniedHttpException(ErrorMessagesEnum::ACCESS_DENIED->value);
+        }
 
         $budget->clearIncomes();
         $budget->clearExpenses();
@@ -56,7 +69,7 @@ class BudgetService
         return $this->createOrUpdateBudget($budgetPayload, $budget);
     }
 
-    private function createOrUpdateBudget(BudgetPayload $budgetPayload, Budget $budget): Budget
+    private function createOrUpdateBudget(BudgetPayload $budgetPayload, Budget $budget): BudgetResponse
     {
         /** @var User $user */
         $user = $this->security->getUser();
@@ -80,19 +93,19 @@ class BudgetService
 
         $this->budgetRepository->save($budget, true);
 
-        return $budget;
+        return $this->createBudgetResponse($budget);
     }
 
-    public function delete(Budget $budget): Budget
+    public function delete(Budget $budget): void
     {
-        $this->checkAccess($budget);
+        if (! $this->authorizationChecker->isGranted(BudgetVoter::DELETE, $budget)) {
+            throw new AccessDeniedHttpException(ErrorMessagesEnum::ACCESS_DENIED->value);
+        }
 
         $this->budgetRepository->delete($budget, true);
-
-        return $budget;
     }
 
-    public function duplicate(?int $id = null): Budget
+    public function duplicate(?int $id = null): BudgetResponse
     {
         if ($id === null) {
             $budget = $this->budgetRepository->findLatestByUser($this->security->getUser());
@@ -101,10 +114,12 @@ class BudgetService
         }
 
         if ($budget === null) {
-            throw new NotFoundHttpException('No budget found');
+            throw new NotFoundHttpException(ErrorMessagesEnum::BUDGET_NOT_FOUND->value);
         }
 
-        $this->checkAccess($budget);
+        if (! $this->authorizationChecker->isGranted(BudgetVoter::VIEW, $budget)) {
+            throw new AccessDeniedHttpException(ErrorMessagesEnum::ACCESS_DENIED->value);
+        }
 
         $newBudget = new Budget();
 
@@ -114,8 +129,10 @@ class BudgetService
         $newBudget->setSavingCapacity($budget->getSavingCapacity());
         $newBudget->setUser($budget->getUser());
 
-        $newDate = $this->budgetRepository->findLatestByUser($budget->getUser())->getDate(); // @phpstan-ignore-line
-        $newDate->modify('+1 month'); // @phpstan-ignore-line
+        $newDate = Carbon::parse($this->budgetRepository->findLatestByUser($budget->getUser())?->getDate())
+            ->startOfMonth()
+            ->addMonth()
+        ;
         $newBudget->setDate($newDate);
 
         foreach ($budget->getIncomes() as $income) {
@@ -132,16 +149,13 @@ class BudgetService
 
         $this->budgetRepository->save($newBudget, true);
 
-        return $newBudget;
+        return $this->createBudgetResponse($newBudget);
     }
 
-    /**
-     * @return SlidingPagination<int, Budget>
-     */
     public function paginate(
         ?PaginationQueryParams $paginationQueryParams = null,
         ?BudgetFilterQuery $budgetFilterQuery = null
-    ): SlidingPagination {
+    ): PaginatedResponseDto {
         $criteria = Criteria::create();
         $criteria->andWhere(Criteria::expr()->eq('user', $this->security->getUser()))
             ->orderBy([
@@ -149,13 +163,55 @@ class BudgetService
             ])
         ;
 
-        return $this->budgetRepository->paginate($paginationQueryParams, $budgetFilterQuery, $criteria);
+        $paginated = $this->budgetRepository->paginate($paginationQueryParams, $budgetFilterQuery, $criteria);
+
+        $budgets = [];
+
+        foreach ($paginated->getItems() as $budget) {
+            /** @var Budget $budget */
+            $budgets[] = $this->createBudgetResponse($budget);
+        }
+
+        return new PaginatedResponseDto(
+            data: $budgets,
+            meta: new PaginationMetaDto(
+                total: $paginated->getTotalItemCount(),
+                page: $paginated->getCurrentPageNumber(),
+                limit: $paginated->getItemNumberPerPage(),
+            ),
+        );
     }
 
-    private function checkAccess(Budget $budget): void
+    private function createBudgetResponse(Budget $budget): BudgetResponse
     {
-        if ($this->security->getUser() !== $budget->getUser()) {
-            throw new AccessDeniedHttpException('Access denied');
+        $incomes = [];
+        foreach ($budget->getIncomes() as $income) {
+            $incomes[] = new IncomeResponse(
+                id: $income->getId(),
+                name: $income->getName(),
+                amount: $income->getAmount()
+            );
         }
+
+        $expenses = [];
+        foreach ($budget->getExpenses() as $expense) {
+            $expenses[] = new ExpenseResponse(
+                id: $expense->getId(),
+                name: $expense->getName(),
+                amount: $expense->getAmount(),
+                category: $expense->getCategory()
+            );
+        }
+
+        return new BudgetResponse(
+            id: $budget->getId(),
+            name: $budget->getName(),
+            incomesAmount: $budget->getIncomesAmount(),
+            expensesAmount: $budget->getExpensesAmount(),
+            savingCapacity: $budget->getSavingCapacity(),
+            date: $budget->getDate()->format('Y-m'),
+            incomes: $incomes,
+            expenses: $expenses
+        );
     }
 }
